requirement.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
codegen==1.0
decorator==4.3.0
Django==1.8.4
django-filter==1.1.0
django-webpack-loader==0.2.4
djangorestframework==3.3.2
enum34==1.1.6
ipython==5.7.0
ipython-genutils==0.2.0
Markdown==2.6.11
numpy==1.14.5
pathlib2==2.3.2
patterns==0.3
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.4
PyYAML==3.12
scandir==1.7
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
